gidari360 remove unecessary data
