--- examples/mongodb/go.mod
@@ -1,27 +1,0 @@
-module github.com/alpstable/gidari/examples/mongodb
-
-go 1.19
-
-replace github.com/alpstable/gidari => ../../
-
-require (
-	github.com/alpstable/gidari v0.0.0-20221106191643-d53cdac2d61f
-	github.com/alpstable/gmongo v0.0.0-20221106194455-977f4e3bb3d6
-	go.mongodb.org/mongo-driver v1.11.0
-)
-
-require (
-	github.com/golang/snappy v0.0.1 // indirect
-	github.com/klauspost/compress v1.13.6 // indirect
-	github.com/montanaflynn/stats v0.0.0-20171201202039-1bf9dbcd8cbe // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/xdg-go/pbkdf2 v1.0.0 // indirect
-	github.com/xdg-go/scram v1.1.1 // indirect
-	github.com/xdg-go/stringprep v1.0.3 // indirect
-	github.com/youmark/pkcs8 v0.0.0-20181117223130-1be2e3e5546d // indirect
-	golang.org/x/crypto v0.0.0-20220622213112-05595931fe9d // indirect
-	golang.org/x/sync v0.1.0 // indirect
-	golang.org/x/text v0.3.7 // indirect
-	golang.org/x/time v0.0.0-20220722155302-e5dcc9cfc0b9 // indirect
-	google.golang.org/protobuf v1.28.1 // indirect
-)
