+++ gidari_example_test.go
@@ -1,55 +1,76 @@
 package gidari_test
 
-//func ExampleNewIterator() {
-//	ctx := context.Background()
+import (
+	"context"
+	"fmt"
+	"io"
+	"log"
+	"net/http"
+	"time"
+
+	"github.com/alpstable/gidari"
+	"golang.org/x/time/rate"
+)
+
+func ExampleHTTPService_Iterator() {
+	ctx := context.Background()
+
+	const api = "https://anapioficeandfire.com/api"
+
+	// First we create a service that can be used to make bulk HTTP
+	// requests to the API.
+	svc, err := gidari.NewService(ctx)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// Create some requests and add them to the service.
+	bookReq, _ := http.NewRequest(http.MethodGet, api+"/books", nil)
+	charReq, _ := http.NewRequest(http.MethodGet, api+"/characters", nil)
+	housReq, _ := http.NewRequest(http.MethodGet, api+"/houses", nil)
+
+	svc.HTTP.
+		Requests(&gidari.HTTPRequest{Request: bookReq}).
+		Requests(&gidari.HTTPRequest{Request: charReq}).
+		Requests(&gidari.HTTPRequest{Request: housReq})
+
+	// Add a rate limiter to the service, 5 requests per second. This can
+	// help avoid "429" errors.
+	svc.HTTP.RateLimiter(rate.NewLimiter(rate.Every(1*time.Second), 5))
+
+	// byteSize will keep track of the sum of bytes for each HTTP Response's
+	// body.
+	var byteSize int
+	for svc.HTTP.Iterator.Next(ctx) {
+		current := svc.HTTP.Iterator.Current
+
+		rsp := current.Response
+		if rsp == nil {
+			break
+		}
+
+		// Get the byte slice from the response body.
+		body, err := io.ReadAll(current.Response.Body)
+		if err != nil {
+			log.Fatalf("failed to read response body: %v", err)
+		}
+
+		// Add the number of bytes to the sum.
+		byteSize += len(body)
+	}
+
+	// Check to see if an error occurred.
+	if err := svc.HTTP.Iterator.Err(); err != nil {
+		panic(err)
+	}
+
+	fmt.Println("Total number of bytes:", byteSize)
+	// Output:
+	// Total number of bytes: 256179
+}
+
 //
-//	const api = "https://anapioficeandfire.com/api"
-//
-//	// Create the HTTP Requests to iterate over.
-//	bookReq, _ := http.NewRequestWithContext(ctx, http.MethodGet, api+"/books", nil)
-//	houseReq, _ := http.NewRequestWithContext(ctx, http.MethodGet, api+"/houses", nil)
-//	charReq, _ := http.NewRequestWithContext(ctx, http.MethodGet, api+"/characters", nil)
-//
-//	// Create the iterator object.
-//	iter, _ := gidari.NewIterator(ctx, &gidari.Config{
-//		Requests: []*gidari.Request{
-//			{
-//				HttpRequest: bookReq,
-//				RateLimiter: rate.NewLimiter(rate.Every(time.Second), 5),
-//			},
-//			{
-//				HttpRequest: houseReq,
-//				RateLimiter: rate.NewLimiter(rate.Every(time.Second), 5),
-//			},
-//			{
-//				HttpRequest: charReq,
-//				RateLimiter: rate.NewLimiter(rate.Every(time.Second), 5),
-//			},
-//		},
-//	})
-//
-//	defer iter.Close(ctx)
-//
-//	// byteSize will keep track of the sum of bytes for each HTTP Response's body.
-//	var byteSize int
-//
-//	for iter.Next(ctx) {
-//		// Get the byte slice from the response body.
-//		body, err := io.ReadAll(iter.Current.Body)
-//		if err != nil {
-//			log.Fatalf("failed to read response body: %v", err)
-//		}
-//
-//		// Add the number of bytes to the sum.
-//		byteSize += len(body)
-//	}
-//
-//	fmt.Println("Total number of bytes:", byteSize)
-//	// Output:
-//	// Total number of bytes: 256179
-//}
-//
-//func ExampleTransport() {
+// func ExampleTransport() {
 //	ctx := context.Background()
 //
 //	const api = "https://anapioficeandfire.com/api"
